model changes for readingFolder And old AssoMailPost

+++ src/main/java/com/jss/osiris/modules/myjss/wordpress/service/ReadingFolderServiceImpl.java
@@ -1,13 +1,12 @@
 package com.jss.osiris.modules.myjss.wordpress.service;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.jss.osiris.modules.myjss.wordpress.model.AssoMailPost;
-import com.jss.osiris.modules.myjss.wordpress.model.Media;
 import com.jss.osiris.modules.myjss.wordpress.model.ReadingFolder;
 import com.jss.osiris.modules.osiris.profile.service.EmployeeService;
 import com.jss.osiris.modules.osiris.quotation.repository.ReadingFolderRepository;
@@ -20,9 +19,6 @@ public class ReadingFolderServiceImpl implements ReadingFolderService {
 
     @Autowired
     EmployeeService employeeService;
-
-    @Autowired
-    AssoMailPostService assoMailPostService;
 
     @Override
     public ReadingFolder addOrUpdateReadingFolder(ReadingFolder readingFolder) {
@@ -53,31 +49,27 @@ public class ReadingFolderServiceImpl implements ReadingFolderService {
 
     @Override
     public void deleteReadingFolder(ReadingFolder readingFolder) {
-        List<AssoMailPost> assoMailPosts = null;
-        if (readingFolder.getId() != null) {
-            assoMailPosts = assoMailPostService.getAssoMailPostsByReadingFolder(readingFolder);
-            if (!assoMailPosts.isEmpty())
-                for (AssoMailPost asso : assoMailPosts) {
-                    asso.setReadingFolder(null);
-                    assoMailPostService.addOrUpdateAssoMailPost(asso);
-                }
-            readingFolderRepository.delete(readingFolder);
+        Responsable responsable = employeeService.getCurrentMyJssUser();
+        if (responsable != null) {
+            if (readingFolder.getId() != null
+                    && readingFolder.getMail().getId().equals(responsable.getMail().getId())) {
+                readingFolderRepository.delete(readingFolder);
+            }
         }
     }
 
     @Override
-    public ReadingFolder getFirstPostImage(ReadingFolder readingFolder) {
-        List<AssoMailPost> assoMailPosts = assoMailPostService.getAssoMailPostsByReadingFolder(readingFolder);
-        Media firstMedia = null;
+    public ReadingFolder initReadingFolderForCurrentUser() {
+        Responsable responsable = employeeService.getCurrentMyJssUser();
+        ReadingFolder readingFolder = null;
 
-        if (!assoMailPosts.isEmpty()) {
-            for (AssoMailPost assoMailPost : assoMailPosts) {
-                if (assoMailPost.getPost() != null && assoMailPost.getPost().getMedia() != null) {
-                    firstMedia = assoMailPost.getPost().getMedia();
-                    readingFolder.setMedia(firstMedia);
-                    addOrUpdateReadingFolder(readingFolder);
-                    break;
-                }
+        if (responsable != null && responsable.getMail() != null) {
+            if (this.getAvailableReadingFolders() == null || this.getAvailableReadingFolders().size() == 0) {
+                readingFolder = new ReadingFolder();
+                readingFolder.setLabel("Tous les articles");
+                readingFolder.setMail(responsable.getMail());
+                readingFolder.setPosts(new ArrayList<>());
+                readingFolder = addOrUpdateReadingFolder(readingFolder);
             }
         }
         return readingFolder;
